[front] refactor(mcp): migrate notion server to new pattern
