+++ front/lib/actions/mcp_internal_actions/constants.ts
@@ -23,6 +23,7 @@ import { FILE_GENERATION_SERVER } from "@app/lib/api/actions/servers/file_genera
 import { GITHUB_SERVER } from "@app/lib/api/actions/servers/github/metadata";
 import { GOOGLE_CALENDAR_SERVER } from "@app/lib/api/actions/servers/google_calendar/metadata";
 import { IMAGE_GENERATION_SERVER } from "@app/lib/api/actions/servers/image_generation/metadata";
+import { NOTION_SERVER } from "@app/lib/api/actions/servers/notion/metadata";
 import { STATUSPAGE_SERVER } from "@app/lib/api/actions/servers/statuspage/metadata";
 import {
   DEEP_DIVE_NAME,
@@ -366,44 +367,10 @@ export const INTERNAL_MCP_SERVERS = {
     allowMultipleInstances: true,
     isRestricted: undefined,
     isPreview: false,
-    tools_stakes: {
-      search: "never_ask",
-      retrieve_page: "never_ask",
-      retrieve_database_schema: "never_ask",
-      retrieve_database_content: "never_ask",
-      query_database: "never_ask",
-      retrieve_block: "never_ask",
-      retrieve_block_children: "never_ask",
-      fetch_comments: "never_ask",
-      list_users: "never_ask",
-      get_about_user: "never_ask",
-
-      create_page: "low",
-      insert_row_into_database: "low",
-      create_database: "low",
-      update_page: "low",
-      add_page_content: "low",
-      create_comment: "low",
-      delete_block: "low",
-      delete_page: "low",
-      update_row_database: "low",
-      update_schema_database: "low",
-    },
     tools_arguments_requiring_approval: undefined,
     tools_retry_policies: undefined,
     timeoutMs: undefined,
-    serverInfo: {
-      name: "notion",
-      version: "1.0.0",
-      description: "Access workspace pages and databases.",
-      authorization: {
-        provider: "notion" as const,
-        supported_use_cases: ["platform_actions", "personal_actions"] as const,
-      },
-      icon: "NotionLogo",
-      documentationUrl: "https://docs.dust.tt/docs/notion-mcp",
-      instructions: null,
-    },
+    metadata: NOTION_SERVER,
   },
   extract_data: {
     id: 12,
